# docs/swagger.yaml
definitions:
  param.ItemReq:
    properties:
      description:
        type: string
      id:
        type: integer
      item_code:
        type: string
      quantity:
        type: integer
    type: object
  param.OrderReq:
    properties:
      customer_name:
        type: string
      items:
        items:
          $ref: '#/definitions/param.ItemReq'
        type: array
      order_at:
        type: string
    type: object
info:
  contact: {}
paths:
  /order:
    post:
      consumes:
      - application/json
      description: Create Order
      parameters:
      - description: body data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/param.OrderReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Create Order
  /order/{id}:
    get:
      consumes:
      - application/json
      description: Get Order
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get Order
  /order/{order_id}:
    delete:
      consumes:
      - application/json
      description: Delete Order
      parameters:
      - description: Order Id
        in: path
        name: order_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Delete Order
    put:
      consumes:
      - application/json
      description: Get Order
      parameters:
      - description: Order Id
        in: path
        name: order_id
        required: true
        type: integer
      - description: body data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/param.OrderReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get Order
swagger: "2.0"